Updated examples to run without pip installing martypy
